requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
distlib==0.3.3
filelock==3.3.0
idna==3.2
loguru==0.5.3
platformdirs==2.4.0
pyTelegramBotAPI==4.2.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
win32-setctime==1.0.3
